Indev 12 Performance enhancements
