+++ Programs/NameGen(f)/app.py
@@ -1,13 +1,21 @@
 #  Copyright (c) 2022. luz5020
 #  Approved for use in Open Source Projects
+import random
 import tkinter as tk
 
-import NameGen as Ng
+import pandas as pd
 
 
-# Methods
-def NameGen():
-    name = Ng.namegen()
+def namegen():
+    col_list = ["year", "name", "percent", "sex"]
+    df = pd.read_csv("Names_Short_F.csv", usecols=col_list)
+    # for x in df.index:
+    # if df.loc[x, "yr"] < 2010:
+    #     df.drop(x, inplace=True)
+    dl = df['name'].tolist()
+    n = random.randint(0, len(dl))
+    name = dl[n]
+    # print(dl[n])
     lbl_name["text"] = name
 
 
@@ -22,7 +30,7 @@ for i in range(2):
 btn_gen = tk.Button(
     master=frm_btn,
     text="Generate",
-    command=NameGen
+    command=namegen
 )
 lbl_name = tk.Label(master=frm_out, text="Generated Name")
 # Placement
